Реализована проверка на отрицательный баланс после покупки

+++ payment-service/src/main/java/ru/yandex/practicum/payment_service/core/service/PaymentService.java
@@ -4,8 +4,10 @@ import reactor.core.publisher.Mono;
 import ru.yandex.practicum.payment.api.model.BalanceResponse;
 import ru.yandex.practicum.payment.api.model.PaymentRequest;
 
+import java.math.BigDecimal;
+
 public interface PaymentService {
    BalanceResponse getBalance();
 
-    Mono<Long> makePayment(Mono<PaymentRequest> paymentRequest);
+    Mono<BigDecimal> makePayment(Mono<PaymentRequest> paymentRequest);
 }
